requirements.txt
asgiref==3.8.1
coverage==7.5.3
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
iniconfig==2.0.0
mccabe==0.7.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
patsy==0.5.6
plotly==5.22.0
plotly-express==0.4.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
ruff==0.4.8
scipy==1.13.1
segno==1.6.1
six==1.16.0
sqlparse==0.5.0
statsmodels==0.14.2
tenacity==8.3.0
tzdata==2024.1
uuid==1.30